+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - Made `JavacParseResult` a generic class.
 - Renamed `parseJavaFile()` to `parseFile();`.
 - Renamed `parseJavaCode()` to `parseCompilationUnit()`.
+- Renamed `parseJavaFileObject()` to `parseCompilationUnit()`.
 - New method `parseTypeDeclaration()`.
+- New method `parseTypeUse()`.
 
 ## 0.2.0 (2024-06-15)